# Engenharia de Software/openapi.yaml
openapi: 3.0.3
info:
  title: GitHub Monitor API
  description: |
    API para monitoramento de repositórios GitHub, registro de eventos e sistema de alertas.
    
    Baseado no projeto GitHub Monitor API.
    
    **Funcionalidades Principais:**
    * Cadastro e gerenciamento de repositórios.
    * Recepção de Webhooks do GitHub (Push, PR, Issues, Releases).
    * Consulta de histórico de eventos.
    * Métricas de repositórios.
    * Gerenciamento de regras de alerta e consulta de incidentes.
    * Autenticação e registro de usuários.
  version: 1.0.0
  contact:
    name: Rodrigo Pereira de Souza
    email: rodrigo@example.com
servers:
  - url: http://localhost:8080
    description: Servidor de Desenvolvimento Local
...
paths:
...
  /api/repositories:
    get:
      tags:
        - Repositories
      summary: Listar repositórios cadastrados
      operationId: listRepositories
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
          description: Número da página (paginação)
        - name: size
          in: query
          schema:
            type: integer
            default: 20
          description: Itens por página
      responses:
        '200':
          description: Lista de repositórios retornada com sucesso
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PageMetadata'
                  - type: object
                    properties:
                      content:
                        type: array
                        items:
                          $ref: '#/components/schemas/Repository'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Repositories
      summary: Cadastrar novo repositório para monitoramento
      operationId: createRepository
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RepositoryInput'
      responses:
        '201':
          description: Repositório cadastrado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Repository'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/repositories/{id}:
    delete:
      tags:
        - Repositories
      summary: Remover um repositório
      operationId: deleteRepository
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Repositório removido com sucesso
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/webhooks/github:
    post:
      tags:
        - Webhooks
      summary: Endpoint para receber Webhooks do GitHub
      description: |
        Endpoint público para recebimento de Webhooks do GitHub.
        Todas as requisições devem conter o header:
        X-Hub-Signature-256 para validação da assinatura.
      security: [] 
      operationId: handleGithubWebhook
      parameters:
        - name: X-GitHub-Event
          in: header
          required: true
          schema:
            type: string
          description: "Tipo do evento (ex: push, pull_request)"
        - name: X-Hub-Signature-256
          in: header
          required: true
          schema:
            type: string
          description: Assinatura HMAC SHA-256 do payload
        - name: X-GitHub-Delivery
          in: header
          required: true
          schema:
            type: string
            format: uuid
          description: ID único de entrega do evento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Payload dinâmico do GitHub
      responses:
        '200':
          description: Evento processado com sucesso
        '202':
          description: Evento aceito para processamento assíncrono
        '400':
          description: Assinatura inválida ou payload malformado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Assinatura HMAC ausente ou incorreta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /api/alerts:
    get:
      tags:
        - Alerts
      summary: Consultar histórico de alertas disparados
      operationId: listAlerts
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [OPEN, RESOLVED]
        - name: severity
          in: query
          schema:
            type: string
            enum: [INFO, WARNING, CRITICAL]
        - name: page
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Histórico de alertas
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PageMetadata'
                  - type: object
                    properties:
                      content:
                        type: array
                        items:
                          $ref: '#/components/schemas/Alert'
  /api/alerts/{id}/resolve:
    post:
      tags:
        - Alerts
      summary: Resolver um alerta
      operationId: resolveAlert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Alerta resolvido com sucesso
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: Token de acesso ausente ou inválido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    BadRequest:
      description: Requisição inválida ou dados malformados
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
    ApiError:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2026-01-24T12:00:00Z"
        status:
          type: integer
          example: 400
        error:
          type: string
          example: "VALIDATION_ERROR"
        message:
          type: string
          example: "Campo repositoryId é obrigatório"
        path:
          type: string
          example: "/api/v1/repositories"
...
    PageMetadata:
      type: object
      properties:
        totalElements:
          type: integer
          description: Total de itens encontrados
        totalPages:
          type: integer
          description: Total de páginas
        page:
          type: integer
          description: Página atual
        size:
          type: integer
          description: Itens por página
    Repository:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "github-monitor-api"
        owner:
          type: string
          example: "rodrigo3301"
        url:
          type: string
          example: "https://github.com/rodrigo3301/github-monitor-api"
        lastSyncedAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [synced, failed, syncing, paused, pending]
          description: Status de sincronização do repositório
        language:
          type: string
          example: "Java"
          description: Linguagem principal do repositório
        createdAt:
          type: string
          format: date-time
    RepositoryInput:
      type: object
      required:
        - owner
        - name
      properties:
        owner:
          type: string
          description: Dono do repositório (usuário ou organização)
        name:
          type: string
          description: Nome do repositório
        webhookSecret:
          type: string
          description: Segredo opcional para configurar webhook manualmente (se não auto-configurado)
...
    Alert:
      type: object
      properties:
        id:
          type: string
          format: uuid
        repositoryId:
          type: string
          format: uuid
        ruleType:
          type: string
          example: "HIGH_FREQUENCY_COMMITS"
        severity:
          type: string
          enum: [INFO, WARNING, CRITICAL]
        message:
          type: string
        branch:
          type: string
          description: Nome da branch relacionada ao alerta (se houver)
        authorLogin:
          type: string
          description: Login do autor do evento que gerou o alerta
        authorAvatarUrl:
          type: string
          description: URL do avatar do autor
        status:
          type: string
          enum: [OPEN, RESOLVED]
        createdAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
...
security:
  - bearerAuth: []