# Github_Monitor/src/main/java/com/githubmonitor/api/dto/RepositoryInput.java
package com.githubmonitor.api.dto;

import jakarta.validation.constraints.NotBlank;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class RepositoryInput {
    @NotBlank(message = "Owner is required")
    private String owner;
    
    @NotBlank(message = "Name is required")
    private String name;
    
    private String webhookSecret;
}